Fix a 32-bit overflow in showdb.c.
FossilOrigin-Name: c4d6d3cd3ccf264212eb4c9fdd0d05239627298709c70da2cbca852d0f24fac7

+++ tool/showdb.c
@@ -27,7 +27,7 @@ typedef sqlite3_uint64 u64;       /* unsigned 64-bit */
 
 
 static struct GlobalData {
-  u32 pagesize;                   /* Size of a database page */
+  i64 pagesize;                   /* Size of a database page */
   int dbfd;                       /* File descriptor for reading the DB */
   u32 mxPage;                     /* Last page number */
   int perLine;                    /* HEX elements to print per line */
@@ -1178,7 +1178,7 @@ int main(int argc, char **argv){
   if( g.pagesize==0 ) g.pagesize = 1024;
   sqlite3_free(zPgSz);
 
-  printf("Pagesize: %d\n", g.pagesize);
+  printf("Pagesize: %d\n", (int)g.pagesize);
   g.mxPage = (u32)((szFile+g.pagesize-1)/g.pagesize);
 
   printf("Available pages: 1..%u\n", g.mxPage);
